+++ src/app/components/list-all-event/list-all-event.component.ts
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import { Component } from '@angular/core';
 import { EventService } from 'src/app/services/event/event.service';
 
@@ -28,15 +29,33 @@ export class ListAllEventComponent {
   }
 
   borrar(id: number) {
-    this.serviceevent.deleteEvent(id).subscribe(
-      data =>{
-        console.log(data);
-        
-        this.Event = data;
-        console.log(this.Event);
-        window.location.reload()
+
+    Swal.fire({
+      title: 'Desea eliminar este registro? ',
+      showCancelButton: true,
+      confirmButtonText: 'Si',
+
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.serviceevent.deleteEvent(id).subscribe(
+          data =>{
+            console.log(data);
+            
+            this.Event = data;
+            console.log(this.Event);
+            window.location.reload()
+          }
+        )
+        Swal.fire('El registro se elimino con exito')
+        setTimeout(() => {
+          
+          window.location.reload()
+         
+        }, 1000);
+ 
       }
-    )
+    })
+   
   }
 
 }
