+++ src/app/components/list-all-responsible/list-all-responsible.component.ts
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import { Component } from '@angular/core';
 import { ResponsibleService } from 'src/app/services/responsible/responsible.service';
 
@@ -28,15 +29,33 @@ export class ListAllResponsibleComponent {
   }
 
   borrar(document: number) {
-    this.serviceresponsible.deleteResponsible(document).subscribe(
-      data =>{
-        console.log(data);
-        
-        this.responsible = data;
-        console.log(this.responsible);
-        window.location.reload()
+
+    Swal.fire({
+      title: 'Desea eliminar este registro? ',
+      showCancelButton: true,
+      confirmButtonText: 'Si',
+
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.serviceresponsible.deleteResponsible(document).subscribe(
+          data =>{
+            console.log(data);
+            
+            this.responsible = data;
+            console.log(this.responsible);
+            window.location.reload()
+          }
+        )
+        Swal.fire('El registro se elimino con exito')
+        setTimeout(() => {
+          
+          window.location.reload()
+         
+        },1000);
+ 
       }
-    )
+    })
+
   }
 
 }
